requirements.txt
customtkinter>=5.2.0
psutil>=5.9.0
wmi>=1.5.1
pillow>=10.0.0
pyinstaller>=6.0.0
